+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [MASTER BRANCH - NOT RELEASED]
+
+### Added
+
+* Add Tracis CI service
+    * Add Travis builds widget
+* Add Feedly service
+    * Add Feedly subscribers widget
+
 ## [0.2.0] - 2019-10-05
 
 ### Added
@@ -26,13 +35,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Replace `title_options` by `name_options` for project's config (breaking change)
 
-## [0.1.1] - 2018-07-21
+## [0.1.1] - 2019-07-21
 
 ### Added
 
 * use goreleaser for relases
 
-## [0.1.0] - 2018-05-28
+## [0.1.0] - 2019-05-28
 
 ### Added